Remove typescript ^4.8 in workspace subdirectories so entire project uses one typescript version (5.3.2)
We came across this while working on https://github.com/ngi-nix/ngipkgs/pull/164.
While not strictly necessary for that task, it seemed like it was unintentional
that the project was depending on different versions of typescript (PR with original
upgrade is https://github.com/atomicdata-dev/atomic-server/pull/710), so
I removed typescript in the various subdirectories so they all depend
on the same version as root.

- Edit tsconfig.build.json to fix the following error
  `TS5110: Option 'module' must be set to 'NodeNext' when option 'moduleResolution' is set to 'NodeNext'.`

- All the removed packages from pnpm-lock.yaml are duplicate entries